[fix] Remove closed producers and consumers from client
Fixes https://github.com/apache/pulsar-client-cpp/issues/55

### Motivation

1. When a producer or consumer is closed, the reference is still stored in `ClientImpl`. If a client kept creating producers or consumers, the memory usage would not reduce.
2. When the `HandlerBase::connection_` field is modified, the `removeProducer` or `removeConsumer` method is not called. Then these producers and consumers will be cached in the connection until the connection is closed.

### Modifications

In `ClientImpl`, use `SynchronizedHashMap` to store references of producers and consumers. The key is the pointer, the value is the `weak_ptr` objects. Then, add `cleanupProducer` and `cleanupConsumer` methods to remove a specific reference for a given pointer.

To call these two methods, this PR does a refactoring that all `ClientImpl` references in producer or consumer are changed to a trivial reference `ClientImpl&` instead of a `weak_ptr`. It's because the lifetime of `ClientImpl` should be longer than them. Using `weak_ptr` is redundant and will increase many repeat codes like:

```c++
auto client = client_.lock();
if (client) {
    client->cleanupProducer(this);
}
```

After that, unify the `shutdown` implementations for producers and consumers, which resets some states, cancels the timers, and unregister itself from both `ClientImpl` and `ClientConnection`. Add adding customized non-null callback in `closeAsync` and `unsubscribeAsync`.

For `ClientConnection`, add a `beforeConnectionChange` virtual method to `HandlerBase`, which is called before `connection_` field is modified. And disallow the direct access to `connection_` from derived classes.

### Verifications

`ClientCloseTest` is added to verify the following cases:
- a single topic
- a partitioned topic (multiple topics)
- a partitioned topic with regex subscription

`ClientCloseTest` is to verify these handlers are removed from `ClientImpl` and `ClientConnection`. `testShutdown` is to verify the handlers will be closed after the owned `Client` instance destroyed.