Fix race condition of hasMessageAvailable by reading latest startMessageId each time
### Motivation

See https://github.com/apache/pulsar-client-python/issues/199

There is a race condition when `hasMessageAvailable` is called after
`seek` if the start message ID of `Reader` is earliest.

In `ConsumerImpl::hasMessageAvailableAsync`, if the connection is not
established at the moment, `lastDequedMessageId_` will be `earliest`
because no message is received. Since `lastMessageIdInBroker_` is also
`earliest`, `getLastMessageIdAsync` will be called and then it comes at

https://github.com/apache/pulsar-client-cpp/blob/e2cacb7dfb57b6d059b49fead2e1611548ff89b0/lib/ConsumerImpl.cc#L1554

However, before `getLastMessageIdAsync` is called, `messageId` was
`earliest` because `lastDequedMessageId_` and `startMessageId_` were
both `earliest`. However, when the callback is called, the
`startMessageId_` has already been updated to `latest` in
`connectionOpened`, so we should compare to `latest`.

### Modifications

In the callback of `getLastMessageIdAsync`, retrieve the latest value of
`startMessageId_` to compare rather then reusing the old value.

Refactor the seek flow to reset the seek states and trigger the callback
after updating the `startMessageId_`.

`ReaderTest.testHasMessageAvailableAfterSeekToEnd` is added to cover the
changes.